add MockSnapshotUtils
