+++ kernel/kernel-api/src/test/scala/io/delta/kernel/test/MockSnapshotUtils.scala
@@ -1,0 +1,73 @@
+package io.delta.kernel.test
+import java.util.Optional
+
+import scala.collection.JavaConverters._
+
+import io.delta.kernel.TransactionSuite.testSchema
+import io.delta.kernel.internal.{SnapshotImpl, TableConfig}
+import io.delta.kernel.internal.actions.{Format, Metadata}
+import io.delta.kernel.internal.fs.Path
+import io.delta.kernel.internal.metrics.SnapshotQueryContext
+import io.delta.kernel.internal.snapshot.LogSegment
+import io.delta.kernel.internal.util.FileNames
+import io.delta.kernel.internal.util.VectorUtils.{buildArrayValue, stringStringMapValue}
+import io.delta.kernel.types.StringType
+import io.delta.kernel.utils.FileStatus
+
+object MockSnapshotUtils {
+
+  /**
+   * Creates a mock snapshot with valid metadata at the given version.
+   * @param ictEnablementInfoOpt Controls the enablement state of in-commit timestamps.
+   */
+  def getMockSnapshot(
+      dataPath: Path,
+      latestVersion: Long,
+      ictEnablementInfoOpt: Option[(Long, Long)] = None): SnapshotImpl = {
+    val configuration = ictEnablementInfoOpt match {
+      case Some((version, _)) if version == 0L =>
+        Map(TableConfig.IN_COMMIT_TIMESTAMPS_ENABLED.getKey -> "true")
+      case Some((version, ts)) =>
+        Map(
+          TableConfig.IN_COMMIT_TIMESTAMPS_ENABLED.getKey -> "true",
+          TableConfig.IN_COMMIT_TIMESTAMP_ENABLEMENT_VERSION.getKey -> version.toString,
+          TableConfig.IN_COMMIT_TIMESTAMP_ENABLEMENT_TIMESTAMP.getKey -> ts.toString)
+      case None =>
+        Map[String, String]()
+    }
+    val metadata = new Metadata(
+      "id",
+      Optional.empty(), /* name */
+      Optional.empty(), /* description */
+      new Format(),
+      testSchema.toJson,
+      testSchema,
+      buildArrayValue(java.util.Arrays.asList("c3"), StringType.STRING),
+      Optional.of(123),
+      stringStringMapValue(configuration.asJava));
+    val logPath = new Path(dataPath, "_delta_log")
+
+    val fs = FileStatus.of(
+      FileNames.deltaFile(logPath, latestVersion),
+      1, /* size */
+      1 /* modificationTime */ )
+    val logSegment = new LogSegment(
+      logPath, /* logPath */
+      latestVersion,
+      Seq(fs).asJava, /* deltas */
+      Seq.empty.asJava, /* compactions */
+      Seq.empty.asJava, /* checkpoints */
+      Optional.empty(), /* lastSeenChecksum */
+      0L /* lastCommitTimestamp */
+    )
+    val snapshotQueryContext = SnapshotQueryContext.forLatestSnapshot(dataPath.toString)
+    new SnapshotImpl(
+      dataPath, /* dataPath */
+      logSegment, /* logSegment */
+      null, /* logReplay */
+      null, /* protocol */
+      metadata,
+      snapshotQueryContext /* snapshotContext */
+    )
+  }
+}
